Fix bug when selecting objects using ctrl-click (command click on mac).
When selecting multiple items using ctrl-click, the selectionChange event emits an event where the oldValue array has 0 items and newValue array has 1 item when selecting the first item. But when selecting the second item, the oldValue array incorrectly has 0 items and newValue incorrectly has 1 item. The expected behavior is that on the second ctrl-selection, the oldValue has 1 item and newValue has 2 items.