extend new object when initiating tiles for a OSM or Bing map sublayer
currently when initiating a tile array for a sublayer the original tile is extended rather than extending a new object, this causes the original tiles src attribute to be overwritten so that you end up with two tile arrays pointing to the same object references. This change creates a new object from the original so that they do not share the same object references.